+++ phoenix-core/src/it/java/org/apache/phoenix/end2end/CDCStreamIT.java
@@ -32,6 +32,8 @@ import org.apache.phoenix.query.QueryServices;
 import org.apache.phoenix.query.QueryServicesOptions;
 import org.apache.phoenix.thirdparty.com.google.common.collect.Maps;
 import org.apache.phoenix.util.CDCUtil;
+import org.apache.phoenix.util.EnvironmentEdgeManager;
+import org.apache.phoenix.util.ManualEnvironmentEdge;
 import org.apache.phoenix.util.ReadOnlyProps;
 import org.apache.phoenix.util.TestUtil;
 import org.junit.Assert;
@@ -60,7 +62,6 @@ import static org.apache.phoenix.query.QueryConstants.CDC_PRE_IMAGE;
 import static org.apache.phoenix.query.QueryConstants.CDC_UPSERT_EVENT_TYPE;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertSame;
 import static org.junit.Assert.assertTrue;
 
 @Category(NeedsOwnMiniClusterTest.class)
@@ -166,6 +167,37 @@ public class CDCStreamIT extends CDCBaseIT {
         createCDC(conn, cdc_sql3, null);
         streamName = getStreamName(conn, tableName, cdcName3);
         assertStreamStatus(conn, tableName, streamName, CDCUtil.CdcStreamStatus.ENABLING);
+    }
+
+    @Test
+    public void testCDCStreamTTL() throws Exception {
+        Connection conn = newConnection();
+        String tableName = generateUniqueName();
+        createTableAndEnableCDC(conn, tableName, true);
+        TestUtil.splitTable(conn, tableName, Bytes.toBytes("m"));
+        String sql = "SELECT PARTITION_END_TIME FROM SYSTEM.CDC_STREAM WHERE TABLE_NAME='" + tableName + "'";
+        ResultSet rs = conn.createStatement().executeQuery(sql);
+        int count = 0;
+        while (rs.next()) {
+            count++;
+        }
+        Assert.assertEquals(3, count);
+        ManualEnvironmentEdge injectEdge = new ManualEnvironmentEdge();
+        long t = System.currentTimeMillis() + QueryServicesOptions.DEFAULT_PHOENIX_CDC_STREAM_PARTITION_EXPIRY_MIN_AGE_MS;
+        t = (t / 1000) * 1000;
+        EnvironmentEdgeManager.injectEdge(injectEdge);
+        injectEdge.setValue(t);
+        rs = conn.createStatement().executeQuery(sql);
+        int newCount = 0;
+        while (rs.next()) {
+            // parent partition row with non-zero end time should have expired
+            if (rs.getLong(1) > 0) {
+                Assert.fail("Closed partition should have expired after TTL.");
+            }
+            newCount++;
+        }
+        Assert.assertEquals(2, newCount);
+        EnvironmentEdgeManager.reset();
     }
 
     /**
